Manual Competitor Finder
Tried to create a database for the competitor finder algorithm. Tried a few different approaches:
- yfinance
- investpy & yfinance
- finnhub free tier API
- Financial Modelling Prep free tier API
- Alpha Vantage free API

Tried all of the above and combinations of them. Had no success in creating a database for the competitor finder algorithm as they all either produced very small datasets (<100 companies) or relatively big datasets ( ~2000 companies) but with a lot of incomplete data. As this approach doesn't work I decided to store all of the python files and collected data in a graveyard folder in case I need them in the future before trying a different approach.